re-organise method
